requirements.txt
aiohttp
asyncpg
discord.py @ git+https://github.com/Rapptz/discord.py@master
jishaku
